Reorg models into blocks. Rename kit to burn_ext.
