+++ crates/bunsen/src/blocks/transformers/mod.rs
@@ -1,3 +1,4 @@
 //! Transformer Components
 pub mod attention;
 pub mod embedding;
+pub mod nanochat;
