+++ crates/bunsen/src/blocks/transformers/nanochat/mod.rs
