+++ crates/bunsen/src/burn_ext/mod.rs
@@ -1,7 +1,7 @@
 //! Burn Extension/Overlay Utilities
 //!
 //! This module contains utilities which are direct extensions
-//! or overlay/replacements for existing kit functionality.
+//! or overlay/replacements for existing `burn_ext` functionality.
 
 pub mod descriptors;
 pub mod distribution;
