+++ crates/bunsen/src/burn_ext/module/mod.rs
@@ -1,4 +1,4 @@
-//! Module support utilities for kit module.
+//! Module support utilities for `burn_ext` module.
 
 #[cfg(feature = "reflection")]
 pub mod reflection;
