+++ crates/bunsen/src/burn_ext/module/reflection/module_visitors/container_type_util.rs
