+++ crates/bunsen/src/burn_ext/optim/lr_selectors.rs
