--- crates/bunsen/src/models/mod.rs
@@ -1,2 +1,0 @@
-//! Larger aggregate models.
-pub mod transformers;
