--- crates/bunsen/src/models/transformers/mod.rs
@@ -1,2 +1,0 @@
-//! Transformer models.
-pub mod nanochat;
